refactor: centralize browser crypto helpers and use typed errors
- Extract base64ToBytes, bytesToBase64, toArrayBuffer, deriveAesKey into
  shared src/lib/browser-crypto.ts module
- Update identity-store.web.ts and identity-backup.web.ts to import from
  the shared module instead of maintaining duplicate implementations
- Replace raw Error throws with SpacesError typed errors in
  identity-backup.web.ts
- Add console.error context logging before throws in security-sensitive
  backup recovery paths
